+++ .github/workflows/deploy-pr.yaml
@@ -42,6 +42,6 @@ jobs:
       - name: Deploy
         run: |
           bash <(curl -fsSL cli.new) && \
-          bunx concurrently "railway up --ci --service website" "railway up --ci --service worker"
+          railway up --ci --service website
         env:
           RAILWAY_TOKEN: ${{ secrets.RAILWAY_TOKEN_PR }}
