+++ .github/workflows/deploy-prod.yaml
@@ -43,6 +43,6 @@ jobs:
       - name: Deploy
         run: |
           bash <(curl -fsSL cli.new) && \
-          bunx concurrently "railway up --ci --service website" "railway up --ci --service worker"
+          railway up --ci --service website
         env:
           RAILWAY_TOKEN: ${{ secrets.RAILWAY_TOKEN_PROD }}
