+++ apps/next/.env.example
@@ -12,10 +12,6 @@ HETZNER_MANAGEMENT_CLUSTER_CA_DATA='x'
 HETZNER_MANAGEMENT_CLUSTER_USER_NAME=x
 HETZNER_MANAGEMENT_CLUSTER_USER_CERT_DATA='x'
 HETZNER_MANAGEMENT_CLUSTER_USER_KEY_DATA='x'
-TEMPORAL_ADDRESS=x
-TEMPORAL_NAMESPACE=x
-TEMPORAL_CLIENT_CERT_DATA='x'
-TEMPORAL_CLIENT_KEY_DATA='x'
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=x
 CLERK_PUBLISHABLE_KEY=x
 CLERK_SECRET_KEY=x
@@ -32,3 +28,6 @@ CLOUD_BUILDER_USERNAME=x
 CLOUD_BUILDER_PASSWORD=x
 CLOUD_BUILDER_NAME=x
 LOOPS_WAITLIST_FORM_URL=x
+INNGEST_BASE_URL=x
+INNGEST_ENV=Development
+INNGEST_SIGNING_KEY=x
