+++ apps/next/.gitignore
@@ -40,7 +40,6 @@ package-lock.json
 
 .nixpacks
 TODO
-temporal/lib
 
 # Deployment files
 temp/
