+++ apps/next/package.json
@@ -1,25 +1,21 @@
 {
-  "name": "metal",
+  "name": "@metal/webapp",
   "version": "0.1.0",
   "engines": {
     "node": "20"
   },
   "private": true,
   "scripts": {
-    "build:website": "next build",
-    "build:worker": "tsc --build ./tsconfig.worker.json",
-    "build": "concurrently --raw \"npm:build:website\" \"npm:build:worker\"",
+    "build": "next build",
     "db:push": "NODE_TLS_REJECT_UNAUTHORIZED=0 drizzle-kit push",
     "db:studio": "NODE_TLS_REJECT_UNAUTHORIZED=0 bunx drizzle-kit studio",
     "db:nuke": "bun run ./app/server/db/nuke.ts",
     "db:psql": "bun exec 'psql $POSTGRES_URL'",
     "dev:website": "next dev --turbo",
     "dev:otel-collector": "cd hack/opentelemetry-collector-dev-setup && docker-compose up > /dev/null 2>&1",
-    "dev:worker": "tsx --tsconfig ./tsconfig.worker.json --env-file .env --watch --watch-preserve-output ./temporal/src/worker.ts",
     "dev": "concurrently --raw \"npm:dev:*\"",
     "lint": "next lint",
-    "start:website": "next start",
-    "start:worker": "bunx tsx --tsconfig ./tsconfig.worker.json ./temporal/src/worker.ts"
+    "start:website": "next start"
   },
   "dependencies": {
     "@clerk/backend": "^1.2.1",
@@ -32,6 +28,10 @@
     "@hono/zod-openapi": "^0.11.0",
     "@joaomoreno/unique-names-generator": "^5.1.0",
     "@kubernetes/client-node": "^0.20.0",
+    "@metal/worker": "workspace:*",
+    "@nivo/bar": "^0.87.0",
+    "@nivo/core": "^0.87.0",
+    "@nivo/line": "^0.87.0",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/sdk-node": "^0.50.0",
     "@opentelemetry/sdk-trace-node": "^1.25.0",
@@ -48,15 +48,12 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
+    "@radix-ui/react-toggle": "^1.1.0",
+    "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@rwh/keystrokes": "^1.5.6",
     "@rwh/react-keystrokes": "^1.5.6",
     "@tanstack/react-table": "^8.16.0",
-    "@temporalio/activity": "^1.9.3",
-    "@temporalio/client": "^1.9.3",
-    "@temporalio/interceptors-opentelemetry": "^1.9.3",
-    "@temporalio/worker": "^1.9.3",
-    "@temporalio/workflow": "^1.9.3",
     "@tsconfig/node20": "^20.1.4",
     "@vercel/otel": "^1.8.3",
     "canvas-confetti": "^1.9.3",
@@ -68,10 +65,12 @@
     "cmdk": "^1.0.0",
     "constructs": "^10.3.0",
     "currency-symbol-map": "^5.1.0",
+    "dayjs": "^1.11.11",
     "dotenv": "^16.4.5",
     "drizzle-orm": "^0.30.10",
     "drizzle-zod": "^0.5.1",
     "hono": "^4.2.5",
+    "inngest": "^3.19.17",
     "js-yaml": "^4.1.0",
     "json-stable-stringify": "^1.1.1",
     "lodash": "^4.17.21",
@@ -88,6 +87,7 @@
     "rimraf": "^5.0.7",
     "sharp": "^0.33.3",
     "sshpk": "^1.18.0",
+    "styled-components": "^6.1.11",
     "tailwind-merge": "^2.2.2",
     "tailwindcss-animate": "^1.0.7",
     "tmp": "^0.2.3",
@@ -113,7 +113,6 @@
     "openapi-typescript": "^6.7.5",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
-    "terser-webpack-plugin": "^5.3.10",
     "typescript": "^5.4.5"
   }
 }
