--- apps/next/tsconfig.worker.json
@@ -1,30 +1,0 @@
-// based on https://learn.temporal.io/tutorials/typescript/nextjs/
-{
-  "extends": "@tsconfig/node20/tsconfig.json",
-  "version": "4.4.2",
-  "compilerOptions": {
-    "emitDecoratorMetadata": false,
-    "experimentalDecorators": false,
-    "declaration": true,
-    "declarationMap": true,
-    "noUncheckedIndexedAccess": true,
-    "sourceMap": true,
-    "composite": true,
-    "outDir": "./temporal/lib",
-    "paths": {
-      "@db/*": ["./app/server/db/*"],
-      "@lib/*": ["./lib/*"]
-    }
-  },
-  "include": [
-    "src/temporal/**/*.ts",
-    "./app/server/db/schema*.ts",
-    "./app/server/db/index.ts",
-    "./lib/constants.ts",
-    "./lib/hcloud/index.d.ts",
-    "./lib/tracedExec.ts",
-    "./lib/k8s.ts",
-    "./types/*"
-  ],
-  "exclude": ["node_modules"]
-}
