+++ packages/worker/types/uuid-base62.d.ts
@@ -1,0 +1,5 @@
+declare module "uuid-base62" {
+  export function encode(uuid: string): string;
+  export function decode(str: string): string;
+  export function v4(): string;
+}
